pyproject.toml
[tool.poetry]
name = "pyray"
version = "0.1.0"
description = ""
authors = ["Alex <alex.mors@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pyside6 = "^6.4.3"
snakeviz = "^2.1.1"
pyglm = "^2.7.0"
numpy = "^1.24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
